[DF-7937] layouts: review fixes
Former-commit-id: 47dab80e1fc19f4597950397aa7f31f064bfa2a1

+++ Sources/Internal/UI/Layouts/FlowLayoutAlgorithm.cpp
@@ -259,9 +259,9 @@ void FlowLayoutAlgorithm::ProcessYAxis(ControlLayoutData &data, UIFlowLayoutComp
 {
     CalculateVerticalDynamicPaddingAndSpaces(data);
 
-    float32 lineHeight = 0;
+    float32 lineHeight = 0.0f;
     float32 y = verticalPadding;
-    float32 firstIndex = data.GetFirstChildIndex();
+    int32 firstIndex = data.GetFirstChildIndex();
     for (int32 index = data.GetFirstChildIndex(); index <= data.GetLastChildIndex(); index++)
     {
         ControlLayoutData &childData = layoutData[index];
